refactor asset drivers
