+++ data/postgres/driver.go
@@ -20,14 +20,11 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/google/uuid"
 	"github.com/jackc/pgconn"
 	"github.com/jackc/pgerrcode"
 	_ "github.com/jackc/pgx/v5/stdlib"
 
 	"arcadium.dev/core/sql"
-
-	"arcadium.dev/arcade/asset"
 )
 
 // Open opens a database.
@@ -46,195 +43,6 @@ func Open(ctx context.Context, dsn string) (*sql.DB, error) {
 
 	return db, nil
 }
-
-type (
-	ItemDriver struct {
-		Driver
-	}
-)
-
-const (
-	ItemListQuery   = `SELECT id, name, description, owner_id, location_item_id, location_player_id, location_room_id, created, updated FROM items`
-	ItemGetQuery    = `SELECT id, name, description, owner_id, location_item_id, location_player_id, location_room_id, created, updated FROM items WHERE id = $1`
-	ItemCreateQuery = `INSERT INTO items (name, description, owner_id, location_item_id, location_player_id, location_room_id) ` +
-		`VALUES ($1, $2, $3, $4, $5, $6) ` +
-		`RETURNING id, name, description, owner_id, location_item_id, location_player_id, location_room_id, created, updated`
-	ItemUpdateQuery = `UPDATE items SET name = $2, description = $3, owner_id = $4, location_item_id = $5, location_player_id = $6, location_room_id = $7 ` +
-		`WHERE id = $1 ` +
-		`RETURNING id, name, description, owner_id, location_item_id, location_player_id, location_room_id, created, updated`
-	ItemRemoveQuery = `DELETE FROM items WHERE id = $1`
-)
-
-// ListQuery returns the List query string given the filter.
-func (ItemDriver) ListQuery(filter asset.ItemFilter) string {
-	fq := ""
-	cmd := "WHERE"
-	if filter.OwnerID != asset.PlayerID(uuid.Nil) {
-		fq += fmt.Sprintf(" %s owner_id = '%s'", cmd, filter.OwnerID)
-		cmd = "AND"
-	}
-	if filter.LocationID != nil {
-		locationID := filter.LocationID.ID()
-		if locationID != asset.LocationID(uuid.Nil) {
-			switch filter.LocationID.Type() {
-			case asset.LocationTypeItem:
-				fq += fmt.Sprintf(" %s location_item_id = '%s'", cmd, locationID)
-			case asset.LocationTypePlayer:
-				fq += fmt.Sprintf(" %s location_player_id = '%s'", cmd, locationID)
-			case asset.LocationTypeRoom:
-				fq += fmt.Sprintf(" %s location_room_id = '%s'", cmd, locationID)
-			}
-		}
-	}
-	fq += limitAndOffset(filter.Limit, filter.Offset)
-	return ItemListQuery + fq
-}
-
-// GetQuery returns the Get query string.
-func (ItemDriver) GetQuery() string { return ItemGetQuery }
-
-// CreateQuery returns the Create query string.
-func (ItemDriver) CreateQuery() string { return ItemCreateQuery }
-
-// UpdateQuery returns the Update query string.
-func (ItemDriver) UpdateQuery() string { return ItemUpdateQuery }
-
-// RemoveQuery returns the Remove query string.
-func (ItemDriver) RemoveQuery() string { return ItemRemoveQuery }
-
-type (
-	LinkDriver struct {
-		Driver
-	}
-)
-
-const (
-	LinkListQuery   = `SELECT id, name, description, owner_id, location_id, destination_id, created, updated FROM links`
-	LinkGetQuery    = `SELECT id, name, description, owner_id, location_id, destination_id, created, updated FROM links WHERE id = $1`
-	LinkCreateQuery = `INSERT INTO links (name, description, owner_id, location_id, destination_id) ` +
-		`VALUES ($1, $2, $3, $4, $5) ` +
-		`RETURNING id, name, description, owner_id, location_id, destination_id, created, updated`
-	LinkUpdateQuery = `UPDATE links SET name = $2, description = $3, owner_id = $4, location_id = $5, destination_id = $6) ` +
-		`WHERE id = $1 ` +
-		`RETURNING id, name, description, owner_id, location_id, destination_id, created, updated`
-	LinkRemoveQuery = `DELETE FROM links WHERE id = $1`
-)
-
-// ListQuery returns the List query string given the filter.
-func (LinkDriver) ListQuery(filter asset.LinkFilter) string {
-	fq := ""
-	cmd := "WHERE"
-	if filter.OwnerID != asset.PlayerID(uuid.Nil) {
-		fq += fmt.Sprintf(" %s owner_id = '%s'", cmd, filter.OwnerID)
-		cmd = "AND"
-	}
-	if filter.LocationID != asset.RoomID(uuid.Nil) {
-		fq += fmt.Sprintf(" %s location_id = '%s'", cmd, filter.LocationID)
-		cmd = "AND"
-	}
-	if filter.DestinationID != asset.RoomID(uuid.Nil) {
-		fq += fmt.Sprintf(" %s destination_id = '%s'", cmd, filter.DestinationID)
-	}
-	fq += limitAndOffset(filter.Limit, filter.Offset)
-	return LinkListQuery + fq
-}
-
-// GetQuery returns the Get query string.
-func (LinkDriver) GetQuery() string { return LinkGetQuery }
-
-// CreateQuery returns the Create query string.
-func (LinkDriver) CreateQuery() string { return LinkCreateQuery }
-
-// UpdateQuery returns the Update query string.
-func (LinkDriver) UpdateQuery() string { return LinkUpdateQuery }
-
-// RemoveQuery returns the Remove query string.
-func (LinkDriver) RemoveQuery() string { return LinkRemoveQuery }
-
-type (
-	PlayerDriver struct {
-		Driver
-	}
-)
-
-const (
-	PlayerListQuery   = `SELECT id, name, description, home_id, location_id, created, updated FROM players`
-	PlayerGetQuery    = `SELECT id, name, description, home_id, location_id, created, updated FROM players WHERE id = $1`
-	PlayerCreateQuery = `INSERT INTO players (name, description, home_id, location_id) ` +
-		`VALUES ($1, $2, $3, $4) ` +
-		`RETURNING id, name, description, home_id, location_id, created, updated`
-	PlayerUpdateQuery = `UPDATE players SET name = $2, description = $3, home_id = $4, location_id = $5 ` +
-		`WHERE id = $1 ` +
-		`RETURNING id, name, description, home_id, location_id, created, updated`
-	PlayerRemoveQuery = `DELETE FROM players WHERE id = $1`
-)
-
-// ListQuery returns the List query string given the filter.
-func (PlayerDriver) ListQuery(filter asset.PlayerFilter) string {
-	fq := ""
-	if filter.LocationID != asset.RoomID(uuid.Nil) {
-		fq += fmt.Sprintf(" WHERE location_id = '%s'", filter.LocationID)
-	}
-	fq += limitAndOffset(filter.Limit, filter.Offset)
-	return PlayerListQuery + fq
-}
-
-// GetQuery returns the Get query string.
-func (PlayerDriver) GetQuery() string { return PlayerGetQuery }
-
-// CreateQuery returns the Create query string.
-func (PlayerDriver) CreateQuery() string { return PlayerCreateQuery }
-
-// UpdateQuery returns the update query string.
-func (PlayerDriver) UpdateQuery() string { return PlayerUpdateQuery }
-
-// RemoveQuery returns the Remove query string.
-func (PlayerDriver) RemoveQuery() string { return PlayerRemoveQuery }
-
-type (
-	RoomDriver struct {
-		Driver
-	}
-)
-
-const (
-	RoomListQuery   = `SELECT id, name, description, owner_id, parent_id, created, updated FROM rooms`
-	RoomGetQuery    = `SELECT id, name, description, owner_id, parent_id, created, updated FROM rooms WHERE id = $1`
-	RoomCreateQuery = `INSERT INTO rooms (name, description, owner_id, parent_id) ` +
-		`VALUES ($1, $2, $3, $4) ` +
-		`RETURNING id, name, description, owner_id, parent_id, created, updated`
-	RoomUpdateQuery = `UPDATE rooms SET name = $2, description = $3, owner_id = $4, parent_id = $5 ` +
-		`WHERE id = $1 ` +
-		`RETURNING id, name, description, owner_id, parent_id, created, updated`
-	RoomRemoveQuery = `DELETE FROM rooms WHERE id = $1`
-)
-
-// ListQuery returns the List query string given the filter.
-func (RoomDriver) ListQuery(filter asset.RoomFilter) string {
-	fq := ""
-	cmd := "WHERE"
-	if filter.OwnerID != asset.PlayerID(uuid.Nil) {
-		fq += fmt.Sprintf(" %s owner_id = '%s'", cmd, filter.OwnerID)
-		cmd = "AND"
-	}
-	if filter.ParentID != asset.RoomID(uuid.Nil) {
-		fq += fmt.Sprintf(" %s parent_id = '%s'", cmd, filter.ParentID)
-	}
-	fq += limitAndOffset(filter.Limit, filter.Offset)
-	return RoomListQuery + fq
-}
-
-// GetQuery returns the Get query string.
-func (RoomDriver) GetQuery() string { return RoomGetQuery }
-
-// CreateQuery returns the Create query string.
-func (RoomDriver) CreateQuery() string { return RoomCreateQuery }
-
-// UpdateQuery returns the Update query string.
-func (RoomDriver) UpdateQuery() string { return RoomUpdateQuery }
-
-// RoomsRemoveQuery returns the Remove query string.
-func (RoomDriver) RemoveQuery() string { return RoomRemoveQuery }
 
 type (
 	Driver struct{}
