+++ data/postgres/item_driver.go
@@ -1,0 +1,78 @@
+//  Copyright 2022-2024 arcadium.dev <info@arcadium.dev>
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package postgres // import "arcadium.dev/arcade/asset/data/postgres"
+
+import (
+	"fmt"
+
+	"github.com/google/uuid"
+
+	"arcadium.dev/arcade/asset"
+)
+
+type (
+	ItemDriver struct {
+		Driver
+	}
+)
+
+const (
+	ItemListQuery   = `SELECT id, name, description, owner_id, location_item_id, location_player_id, location_room_id, created, updated FROM items`
+	ItemGetQuery    = `SELECT id, name, description, owner_id, location_item_id, location_player_id, location_room_id, created, updated FROM items WHERE id = $1`
+	ItemCreateQuery = `INSERT INTO items (name, description, owner_id, location_item_id, location_player_id, location_room_id) ` +
+		`VALUES ($1, $2, $3, $4, $5, $6) ` +
+		`RETURNING id, name, description, owner_id, location_item_id, location_player_id, location_room_id, created, updated`
+	ItemUpdateQuery = `UPDATE items SET name = $2, description = $3, owner_id = $4, location_item_id = $5, location_player_id = $6, location_room_id = $7 ` +
+		`WHERE id = $1 ` +
+		`RETURNING id, name, description, owner_id, location_item_id, location_player_id, location_room_id, created, updated`
+	ItemRemoveQuery = `DELETE FROM items WHERE id = $1`
+)
+
+// ListQuery returns the List query string given the filter.
+func (ItemDriver) ListQuery(filter asset.ItemFilter) string {
+	fq := ""
+	cmd := "WHERE"
+	if filter.OwnerID != asset.PlayerID(uuid.Nil) {
+		fq += fmt.Sprintf(" %s owner_id = '%s'", cmd, filter.OwnerID)
+		cmd = "AND"
+	}
+	if filter.LocationID != nil {
+		locationID := filter.LocationID.ID()
+		if locationID != asset.LocationID(uuid.Nil) {
+			switch filter.LocationID.Type() {
+			case asset.LocationTypeItem:
+				fq += fmt.Sprintf(" %s location_item_id = '%s'", cmd, locationID)
+			case asset.LocationTypePlayer:
+				fq += fmt.Sprintf(" %s location_player_id = '%s'", cmd, locationID)
+			case asset.LocationTypeRoom:
+				fq += fmt.Sprintf(" %s location_room_id = '%s'", cmd, locationID)
+			}
+		}
+	}
+	fq += limitAndOffset(filter.Limit, filter.Offset)
+	return ItemListQuery + fq
+}
+
+// GetQuery returns the Get query string.
+func (ItemDriver) GetQuery() string { return ItemGetQuery }
+
+// CreateQuery returns the Create query string.
+func (ItemDriver) CreateQuery() string { return ItemCreateQuery }
+
+// UpdateQuery returns the Update query string.
+func (ItemDriver) UpdateQuery() string { return ItemUpdateQuery }
+
+// RemoveQuery returns the Remove query string.
+func (ItemDriver) RemoveQuery() string { return ItemRemoveQuery }
